refactor: move moneyString to its own file
